requirements.txt
# core
selenium>=4.21.0
webdriver-manager>=4.0.0
pandas>=2.2
tqdm>=4.66
python-dotenv>=1.0
openpyxl>=3.1
requests>=2.32
browser-cookie3>=0.19.1
pywin32>=306

# typing & linting (optional in prod)
pydantic>=2
